requirements.txt
django
psycopg2
pillow
python-decouple
django-crispy-forms
pytailwindcss
django-modeltranslation